test: split unit and integration CI scopes
Add dedicated Rust unit-test jobs and move proxy database coverage into an integration target so CI can exercise a stable integration subset without mixing scopes.
