+++ src-tauri/src/database/dao/proxy.rs
@@ -981,224 +981,3 @@ impl Database {
         Ok(())
     }
 }
-
-#[cfg(test)]
-mod tests {
-    use crate::database::dao::proxy::PROXY_PREFERENCES_KEY;
-    use crate::database::Database;
-    use crate::error::AppError;
-    use crate::provider::Provider;
-    use serde_json::json;
-
-    fn save_queue_provider(db: &Database, app_type: &str, id: &str) -> Result<(), AppError> {
-        let provider = Provider::with_id(
-            id.to_string(),
-            id.to_string(),
-            json!({"env": {"BASE_URL": "https://example.com"}}),
-            None,
-        );
-        db.save_provider(app_type, &provider)?;
-        db.add_to_failover_queue(app_type, id)?;
-        Ok(())
-    }
-
-    #[tokio::test]
-    async fn test_default_cost_multiplier_round_trip() -> Result<(), AppError> {
-        let db = Database::memory()?;
-
-        let default = db.get_default_cost_multiplier("claude").await?;
-        assert_eq!(default, "1");
-
-        db.set_default_cost_multiplier("claude", "1.5").await?;
-        let updated = db.get_default_cost_multiplier("claude").await?;
-        assert_eq!(updated, "1.5");
-
-        Ok(())
-    }
-
-    #[tokio::test]
-    async fn test_default_cost_multiplier_validation() -> Result<(), AppError> {
-        let db = Database::memory()?;
-
-        let err = db
-            .set_default_cost_multiplier("claude", "not-a-number")
-            .await
-            .unwrap_err();
-        // AppError::localized returns AppError::Localized variant
-        assert!(matches!(
-            err,
-            AppError::Localized {
-                key: "error.invalidMultiplier",
-                ..
-            }
-        ));
-
-        Ok(())
-    }
-
-    #[tokio::test]
-    async fn test_pricing_model_source_round_trip_and_validation() -> Result<(), AppError> {
-        let db = Database::memory()?;
-
-        let default = db.get_pricing_model_source("claude").await?;
-        assert_eq!(default, "response");
-
-        db.set_pricing_model_source("claude", "request").await?;
-        let updated = db.get_pricing_model_source("claude").await?;
-        assert_eq!(updated, "request");
-
-        let err = db
-            .set_pricing_model_source("claude", "invalid")
-            .await
-            .unwrap_err();
-        // AppError::localized returns AppError::Localized variant
-        assert!(matches!(
-            err,
-            AppError::Localized {
-                key: "error.invalidPricingMode",
-                ..
-            }
-        ));
-
-        Ok(())
-    }
-
-    #[tokio::test]
-    async fn clears_supported_failover_rows() -> Result<(), AppError> {
-        let db = Database::memory()?;
-        save_queue_provider(&db, "claude", "claude-p1")?;
-        save_queue_provider(&db, "codex", "codex-p1")?;
-        save_queue_provider(&db, "gemini", "gemini-p1")?;
-        db.set_proxy_flags_sync("claude", true, true)?;
-        db.set_proxy_flags_sync("codex", true, true)?;
-        db.set_proxy_flags_sync("gemini", true, true)?;
-
-        let cleared = db.clear_auto_failover_for_supported_apps().await?;
-
-        assert_eq!(cleared, 3);
-        assert_eq!(db.get_proxy_flags_sync("claude"), (true, false));
-        assert_eq!(db.get_proxy_flags_sync("codex"), (true, false));
-        assert_eq!(db.get_proxy_flags_sync("gemini"), (true, false));
-        Ok(())
-    }
-
-    #[tokio::test]
-    async fn disabling_global_proxy_config_clears_supported_failover_rows() -> Result<(), AppError>
-    {
-        let db = Database::memory()?;
-        save_queue_provider(&db, "claude", "claude-p1")?;
-        save_queue_provider(&db, "codex", "codex-p1")?;
-        save_queue_provider(&db, "gemini", "gemini-p1")?;
-        db.set_proxy_flags_sync("claude", true, true)?;
-        db.set_proxy_flags_sync("codex", true, true)?;
-        db.set_proxy_flags_sync("gemini", true, true)?;
-
-        let mut config = db.get_global_proxy_config().await?;
-        config.proxy_enabled = false;
-        db.update_global_proxy_config(config).await?;
-
-        assert_eq!(db.get_proxy_flags_sync("claude"), (true, false));
-        assert_eq!(db.get_proxy_flags_sync("codex"), (true, false));
-        assert_eq!(db.get_proxy_flags_sync("gemini"), (true, false));
-        Ok(())
-    }
-
-    #[test]
-    fn proxy_preferences_persist_preferred_ports_in_settings_kv() -> Result<(), AppError> {
-        let db = Database::memory()?;
-
-        db.set_app_proxy_preferred_port("codex", 17022)?;
-        db.set_app_proxy_preferred_port("gemini", 17023)?;
-
-        let raw = db
-            .get_setting(PROXY_PREFERENCES_KEY)?
-            .expect("proxy preferences should be stored in settings");
-        assert!(raw.contains("\"preferredPort\":17022"));
-
-        let preferences = db.get_proxy_preferences()?;
-        assert_eq!(
-            preferences
-                .apps
-                .get("codex")
-                .and_then(|preference| preference.preferred_port),
-            Some(17022)
-        );
-
-        Ok(())
-    }
-
-    #[tokio::test]
-    async fn app_preferred_port_falls_back_to_legacy_proxy_config() -> Result<(), AppError> {
-        let db = Database::memory()?;
-        let mut config = db.get_proxy_config().await?;
-        config.listen_port = 17021;
-        db.update_proxy_config(config).await?;
-
-        assert_eq!(db.get_app_proxy_preferred_port("claude")?, 17021);
-
-        db.set_app_proxy_preferred_port("claude", 17022)?;
-        assert_eq!(db.get_app_proxy_preferred_port("claude")?, 17022);
-        Ok(())
-    }
-
-    #[test]
-    fn set_proxy_flags_sync_masks_failover_without_takeover() -> Result<(), AppError> {
-        let db = Database::memory()?;
-
-        db.set_proxy_flags_sync("claude", false, true)?;
-
-        assert_eq!(db.get_proxy_flags_sync("claude"), (false, false));
-        Ok(())
-    }
-
-    #[test]
-    fn set_proxy_flags_sync_masks_failover_with_empty_queue() -> Result<(), AppError> {
-        let db = Database::memory()?;
-
-        db.set_proxy_flags_sync("claude", true, true)?;
-
-        assert_eq!(db.get_proxy_flags_sync("claude"), (true, false));
-        Ok(())
-    }
-
-    #[tokio::test]
-    async fn update_proxy_config_for_app_masks_failover_without_takeover() -> Result<(), AppError> {
-        let db = Database::memory()?;
-        let mut config = db.get_proxy_config_for_app("claude").await?;
-        config.enabled = false;
-        config.auto_failover_enabled = true;
-
-        db.update_proxy_config_for_app(config).await?;
-
-        assert_eq!(db.get_proxy_flags_sync("claude"), (false, false));
-        Ok(())
-    }
-
-    #[tokio::test]
-    async fn update_proxy_config_for_app_masks_failover_with_empty_queue() -> Result<(), AppError> {
-        let db = Database::memory()?;
-        let mut config = db.get_proxy_config_for_app("claude").await?;
-        config.enabled = true;
-        config.auto_failover_enabled = true;
-
-        db.update_proxy_config_for_app(config).await?;
-
-        assert_eq!(db.get_proxy_flags_sync("claude"), (true, false));
-        Ok(())
-    }
-
-    #[tokio::test]
-    async fn update_proxy_config_for_app_preserves_failover_with_non_empty_queue(
-    ) -> Result<(), AppError> {
-        let db = Database::memory()?;
-        save_queue_provider(&db, "claude", "claude-p1")?;
-        let mut config = db.get_proxy_config_for_app("claude").await?;
-        config.enabled = true;
-        config.auto_failover_enabled = true;
-
-        db.update_proxy_config_for_app(config).await?;
-
-        assert_eq!(db.get_proxy_flags_sync("claude"), (true, true));
-        Ok(())
-    }
-}
